This is one part of a larger module that does an assortment of worksheet manipulation.
Note: This code was built in the window os environment and may not work in Mac Excel.
This is one large module that does an assortment of worksheet manipulation. The goal was to break it up into pieces when code fixes or updates were needed. I use some of the standard aspect of the excel macro and later manipulate it for different causes.